% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{AircraftData}
\alias{AircraftData}
\title{QPX Express API Objects 
Finds the least expensive flights between an origin and a destination.}
\usage{
AircraftData(code = NULL, name = NULL)
}
\arguments{
\item{code}{The aircraft code}

\item{name}{The name of an aircraft, for example Boeing 777}
}
\value{
AircraftData object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:08:31
filename: /Users/mark/dev/R/autoGoogleAPI/googleqpxExpressv1.auto/R/qpxExpress_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AircraftData Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The make, model, and type of an aircraft.
}
