% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{AirportData}
\alias{AirportData}
\title{AirportData Object}
\usage{
AirportData(city = NULL, code = NULL, name = NULL)
}
\arguments{
\item{city}{The city code an airport is located in}

\item{code}{An airport's code}

\item{name}{The name of an airport}
}
\value{
AirportData object
}
\description{
AirportData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An airport.
}
