% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{BagDescriptor}
\alias{BagDescriptor}
\title{BagDescriptor Object}
\usage{
BagDescriptor(commercialName = NULL, count = NULL, description = NULL,
  subcode = NULL)
}
\arguments{
\item{commercialName}{Provides the commercial name for an optional service}

\item{count}{How many of this type of bag will be checked on this flight}

\item{description}{A description of the baggage}

\item{subcode}{The standard IATA subcode used to identify this optional service}
}
\value{
BagDescriptor object
}
\description{
BagDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an item of baggage.
}
