% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{CarrierData}
\alias{CarrierData}
\title{CarrierData Object}
\usage{
CarrierData(code = NULL, name = NULL)
}
\arguments{
\item{code}{The IATA designator of a carrier (airline, etc)}

\item{name}{The long, full name of a carrier}
}
\value{
CarrierData object
}
\description{
CarrierData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a carrier (ie. an airline, bus line, railroad, etc) that might be useful to display to an end-user.
}
