% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{CityData}
\alias{CityData}
\title{CityData Object}
\usage{
CityData(code = NULL, country = NULL, name = NULL)
}
\arguments{
\item{code}{The IATA character ID of a city}

\item{country}{The two-character country code of the country the city is located in}

\item{name}{The full name of a city}
}
\value{
CityData object
}
\description{
CityData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a city that might be useful to an end-user; typically the city of an airport.
}

