% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{Data}
\alias{Data}
\title{Data Object}
\usage{
Data(aircraft = NULL, airport = NULL, carrier = NULL, city = NULL,
  tax = NULL)
}
\arguments{
\item{aircraft}{The aircraft that is flying between an origin and destination}

\item{airport}{The airport of an origin or destination}

\item{carrier}{The airline carrier of the aircraft flying between an origin and destination}

\item{city}{The city that is either the origin or destination of part of a trip}

\item{tax}{The taxes due for flying between an origin and a destination}
}
\value{
Data object
}
\description{
Data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detailed information about components found in the solutions of this response, including a trip's airport, city, taxes, airline, and aircraft.
}
