% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{FareInfo}
\alias{FareInfo}
\title{FareInfo Object}
\usage{
FareInfo(basisCode = NULL, carrier = NULL, destination = NULL,
  id = NULL, origin = NULL, private = NULL)
}
\arguments{
\item{basisCode}{No description}

\item{carrier}{The carrier of the aircraft or other vehicle commuting between two points}

\item{destination}{The city code of the city the trip ends at}

\item{id}{A unique identifier of the fare}

\item{origin}{The city code of the city the trip begins at}

\item{private}{Whether this is a private fare, for example one offered only to select customers rather than the general public}
}
\value{
FareInfo object
}
\description{
FareInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Complete information about a fare used in the solution to a low-fare search query. In the airline industry a fare is a price an airline charges for one-way travel between two points. A fare typically contains a carrier code, two city codes, a price, and a fare basis. (A fare basis is a one-to-eight character alphanumeric code used to identify a fare.)
}
