% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{FlightInfo}
\alias{FlightInfo}
\title{FlightInfo Object}
\usage{
FlightInfo(carrier = NULL, number = NULL)
}
\arguments{
\item{carrier}{No description}

\item{number}{The flight number}
}
\value{
FlightInfo object
}
\description{
FlightInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A flight is a sequence of legs with the same airline carrier and flight number. (A leg is the smallest unit of travel, in the case of a flight a takeoff immediately followed by a landing at two set points on a particular carrier with a particular flight number.) The naive view is that a flight is scheduled travel of an aircraft between two points, with possibly intermediate stops, but carriers will frequently list flights that require a change of aircraft between legs.
}
