% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{FreeBaggageAllowance}
\alias{FreeBaggageAllowance}
\title{FreeBaggageAllowance Object}
\usage{
FreeBaggageAllowance(bagDescriptor = NULL, kilos = NULL,
  kilosPerPiece = NULL, pieces = NULL, pounds = NULL)
}
\arguments{
\item{bagDescriptor}{A representation of a type of bag, such as an ATPCo subcode, Commercial Name, or other description}

\item{kilos}{The maximum number of kilos all the free baggage together may weigh}

\item{kilosPerPiece}{The maximum number of kilos any one piece of baggage may weigh}

\item{pieces}{The number of free pieces of baggage allowed}

\item{pounds}{The number of pounds of free baggage allowed}
}
\value{
FreeBaggageAllowance object
}
\description{
FreeBaggageAllowance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about free baggage allowed on one segment of a trip.
}

