% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{LegInfo}
\alias{LegInfo}
\title{LegInfo Object}
\usage{
LegInfo(aircraft = NULL, arrivalTime = NULL, changePlane = NULL,
  connectionDuration = NULL, departureTime = NULL, destination = NULL,
  destinationTerminal = NULL, duration = NULL, id = NULL, meal = NULL,
  mileage = NULL, onTimePerformance = NULL, operatingDisclosure = NULL,
  origin = NULL, originTerminal = NULL, secure = NULL)
}
\arguments{
\item{aircraft}{The aircraft (or bus, ferry, railcar, etc) travelling between the two points of this leg}

\item{arrivalTime}{The scheduled time of arrival at the destination of the leg, local to the point of arrival}

\item{changePlane}{Whether you have to change planes following this leg}

\item{connectionDuration}{Duration of a connection following this leg, in minutes}

\item{departureTime}{The scheduled departure time of the leg, local to the point of departure}

\item{destination}{The leg destination as a city and airport}

\item{destinationTerminal}{The terminal the flight is scheduled to arrive at}

\item{duration}{The scheduled travelling time from the origin to the destination}

\item{id}{An identifier that uniquely identifies this leg in the solution}

\item{meal}{A simple, general description of the meal(s) served on the flight, for example: 'Hot meal'}

\item{mileage}{The number of miles in this leg}

\item{onTimePerformance}{In percent, the published on time performance on this leg}

\item{operatingDisclosure}{Department of Transportation disclosure information on the actual operator of a flight in a code share}

\item{origin}{The leg origin as a city and airport}

\item{originTerminal}{The terminal the flight is scheduled to depart from}

\item{secure}{Whether passenger information must be furnished to the United States Transportation Security Administration (TSA) prior to departure}
}
\value{
LegInfo object
}
\description{
LegInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a leg. (A leg is the smallest unit of travel, in the case of a flight a takeoff immediately followed by a landing at two set points on a particular carrier with a particular flight number.)
}

