% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{PassengerCounts}
\alias{PassengerCounts}
\title{PassengerCounts Object}
\usage{
PassengerCounts(adultCount = NULL, childCount = NULL,
  infantInLapCount = NULL, infantInSeatCount = NULL, seniorCount = NULL)
}
\arguments{
\item{adultCount}{The number of passengers that are adults}

\item{childCount}{The number of passengers that are children}

\item{infantInLapCount}{The number of passengers that are infants travelling in the lap of an adult}

\item{infantInSeatCount}{The number of passengers that are infants each assigned a seat}

\item{seniorCount}{The number of passengers that are senior citizens}
}
\value{
PassengerCounts object
}
\description{
PassengerCounts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The number and type of passengers. Unfortunately the definition of an infant, child, adult, and senior citizen varies across carriers and reservation systems.
}
