% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{PricingInfo}
\alias{PricingInfo}
\title{PricingInfo Object}
\usage{
PricingInfo(baseFareTotal = NULL, fare = NULL, fareCalculation = NULL,
  latestTicketingTime = NULL, passengers = NULL, ptc = NULL,
  refundable = NULL, saleFareTotal = NULL, saleTaxTotal = NULL,
  saleTotal = NULL, segmentPricing = NULL, tax = NULL)
}
\arguments{
\item{baseFareTotal}{The total fare in the base fare currency (the currency of the country of origin)}

\item{fare}{The fare used to price one or more segments}

\item{fareCalculation}{The horizontal fare calculation}

\item{latestTicketingTime}{The latest ticketing time for this pricing assuming the reservation occurs at ticketing time and there is no change in fares/rules}

\item{passengers}{The number of passengers to which this price applies}

\item{ptc}{The passenger type code for this pricing}

\item{refundable}{Whether the fares on this pricing are refundable}

\item{saleFareTotal}{The total fare in the sale or equivalent currency}

\item{saleTaxTotal}{The taxes in the sale or equivalent currency}

\item{saleTotal}{Total per-passenger price (fare and tax) in the sale or equivalent currency}

\item{segmentPricing}{The per-segment price and baggage information}

\item{tax}{The taxes used to calculate the tax total per ticket}
}
\value{
PricingInfo object
}
\description{
PricingInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The price of one or more travel segments. The currency used to purchase tickets is usually determined by the sale/ticketing city or the sale/ticketing country, unless none are specified, in which case it defaults to that of the journey origin country.
}

