% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_functions.R
\docType{package}
\name{qpxExpress_googleAuthR}
\alias{qpxExpress_googleAuthR}
\alias{qpxExpress_googleAuthR-package}
\title{QPX Express API
Finds the least expensive flights between an origin and a destination.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:08:31
filename: /Users/mark/dev/R/autoGoogleAPI/googleqpxExpressv1.auto/R/qpxExpress_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
