% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{SegmentInfo}
\alias{SegmentInfo}
\title{SegmentInfo Object}
\usage{
SegmentInfo(bookingCode = NULL, bookingCodeCount = NULL, cabin = NULL,
  connectionDuration = NULL, duration = NULL, flight = NULL, id = NULL,
  leg = NULL, marriedSegmentGroup = NULL,
  subjectToGovernmentApproval = NULL)
}
\arguments{
\item{bookingCode}{The booking code or class for this segment}

\item{bookingCodeCount}{The number of seats available in this booking code on this segment}

\item{cabin}{The cabin booked for this segment}

\item{connectionDuration}{In minutes, the duration of the connection following this segment}

\item{duration}{The duration of the flight segment in minutes}

\item{flight}{The flight this is a segment of}

\item{id}{An id uniquely identifying the segment in the solution}

\item{leg}{The legs composing this segment}

\item{marriedSegmentGroup}{The solution-based index of a segment in a married segment group}

\item{subjectToGovernmentApproval}{Whether the operation of this segment remains subject to government approval}
}
\value{
SegmentInfo object
}
\description{
SegmentInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of a segment of a flight; a segment is one or more consecutive legs on the same flight. For example a hypothetical flight ZZ001, from DFW to OGG, would have one segment with two legs: DFW to HNL (leg 1), HNL to OGG (leg 2), and DFW to OGG (legs 1 and 2).
}
