% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{SegmentPricing}
\alias{SegmentPricing}
\title{SegmentPricing Object}
\usage{
SegmentPricing(fareId = NULL, freeBaggageOption = NULL, segmentId = NULL)
}
\arguments{
\item{fareId}{A segment identifier unique within a single solution}

\item{freeBaggageOption}{Details of the free baggage allowance on this segment}

\item{segmentId}{Unique identifier in the response of this segment}
}
\value{
SegmentPricing object
}
\description{
SegmentPricing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The price of this segment.
}

