% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TaxData}
\alias{TaxData}
\title{TaxData Object}
\usage{
TaxData(id = NULL, name = NULL)
}
\arguments{
\item{id}{An identifier uniquely identifying a tax in a response}

\item{name}{The name of a tax}
}
\value{
TaxData object
}
\description{
TaxData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tax data.
}

