% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TaxInfo}
\alias{TaxInfo}
\title{TaxInfo Object}
\usage{
TaxInfo(chargeType = NULL, code = NULL, country = NULL, id = NULL,
  salePrice = NULL)
}
\arguments{
\item{chargeType}{Whether this is a government charge or a carrier surcharge}

\item{code}{The code to enter in the ticket's tax box}

\item{country}{For government charges, the country levying the charge}

\item{id}{Identifier uniquely identifying this tax in a response}

\item{salePrice}{The price of the tax in the sales or equivalent currency}
}
\value{
TaxInfo object
}
\description{
TaxInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tax information.
}

