% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TimeOfDayRange}
\alias{TimeOfDayRange}
\title{TimeOfDayRange Object}
\usage{
TimeOfDayRange(earliestTime = NULL, latestTime = NULL)
}
\arguments{
\item{earliestTime}{The earliest time of day in HH:MM format}

\item{latestTime}{The latest time of day in HH:MM format}
}
\value{
TimeOfDayRange object
}
\description{
TimeOfDayRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Two times in a single day defining a time range.
}

