% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TripOption}
\alias{TripOption}
\title{TripOption Object}
\usage{
TripOption(id = NULL, pricing = NULL, saleTotal = NULL, slice = NULL)
}
\arguments{
\item{id}{Identifier uniquely identifying this trip in a response}

\item{pricing}{Per passenger pricing information}

\item{saleTotal}{The total price for all passengers on the trip, in the form of a currency followed by an amount, e}

\item{slice}{The slices that make up this trip's itinerary}
}
\value{
TripOption object
}
\description{
TripOption Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Trip information.
}

