% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TripOptionsRequest}
\alias{TripOptionsRequest}
\title{TripOptionsRequest Object}
\usage{
TripOptionsRequest(maxPrice = NULL, passengers = NULL, refundable = NULL,
  saleCountry = NULL, slice = NULL, solutions = NULL,
  ticketingCountry = NULL)
}
\arguments{
\item{maxPrice}{Do not return solutions that cost more than this price}

\item{passengers}{Counts for each passenger type in the request}

\item{refundable}{Return only solutions with refundable fares}

\item{saleCountry}{IATA country code representing the point of sale}

\item{slice}{The slices that make up the itinerary of this trip}

\item{solutions}{The number of solutions to return, maximum 500}

\item{ticketingCountry}{IATA country code representing the point of ticketing}
}
\value{
TripOptionsRequest object
}
\description{
TripOptionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A QPX Express search request, which will yield one or more solutions.
}

