% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TripOptionsResponse}
\alias{TripOptionsResponse}
\title{TripOptionsResponse Object}
\usage{
TripOptionsResponse(data = NULL, requestId = NULL, tripOption = NULL)
}
\arguments{
\item{data}{Informational data global to list of solutions}

\item{requestId}{An identifier uniquely identifying this response}

\item{tripOption}{A list of priced itinerary solutions to the QPX Express query}
}
\value{
TripOptionsResponse object
}
\description{
TripOptionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A QPX Express search response.
}
