% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_functions.R
\name{trips.search}
\alias{trips.search}
\title{Returns a list of flights.}
\usage{
trips.search(TripsSearchRequest)
}
\arguments{
\item{TripsSearchRequest}{The \link{TripsSearchRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/qpx-express}{Google Documentation}

Other TripsSearchRequest functions: \code{\link{TripsSearchRequest}}
}
