% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TripsSearchRequest}
\alias{TripsSearchRequest}
\title{TripsSearchRequest Object}
\usage{
TripsSearchRequest(request = NULL)
}
\arguments{
\item{request}{A QPX Express search request}
}
\value{
TripsSearchRequest object
}
\description{
TripsSearchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A QPX Express search request.
}
\seealso{
Other TripsSearchRequest functions: \code{\link{trips.search}}
}

