% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpxExpress_objects.R
\name{TripsSearchResponse}
\alias{TripsSearchResponse}
\title{TripsSearchResponse Object}
\usage{
TripsSearchResponse(trips = NULL)
}
\arguments{
\item{trips}{All possible solutions to the QPX Express search request}
}
\value{
TripsSearchResponse object
}
\description{
TripsSearchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A QPX Express search response.
}
